refactor(main): openBrowser を services/browser.ts へ抽出
main プロセス内部(services/packages.ts)からも対話的ダウンロードを
呼べるようにするため、OPEN_BROWSER ハンドラの実装を service 関数化する。
挙動は不変。
